Add false positives and negatives
Also switched 'int' to 'short int' in the definition of the cell structure

+++ src/setup.h
@@ -26,7 +26,7 @@ struct specie {
 
 //A cell is specified by its coordinates and specie type
 struct cell {
-    int x,y,z;
+    short int x,y,z;
     int id;
 };
